Add a default name for subscribers
Refactored the form away from having a hidden field set the name of the
subscriber. All subscribers now get set with the name "Subscriber".
Ideally, there would not be a name set at all but Coherence kicks and
screams when I try to do that and I don't have the time to dig into it
and find how I can change that behavoir.