Add redirect if not signed in
Added an if statement to check if there is a signed in user. If a user
is not signed in, they will be redirected to the sign up page. If a user
is signed in, they would be an admin and will be shown the admin
dashboard. This will require a further check to be added to ensure a
non-admin isn't signed in and trying to access the admin dashboard.