Update forms to have links around needed routes
On the sign up form, updated the heading to reflect that you are signing
up for a newsletter. Added a link to the admin login page. Added a link
for a user to have their email confirmation resent.

On the admin login page, updated the links at the bottom to only be one
link to the newsletter sign up page.