Update mail delivery to be async
Updated the delivery of the newsletter to be async amongst all of the
confirmed subscribed users. It is doing it in a loop but hard to say
what effect a large number of subscribers would have on this.

That being said, if you were dealing with a large number of subscribers,
you would be using something like Drip, SendGrid, etc to gather those
emails and send the emails from within those systems.

Next up will be to change to adapter to the local test adapter, add a
ridiculous amount of users using seeds, and see if I can get this to
blow up.