Add test for a signed in user
A user should not be able to sign in. This test will be rendered
obsolete after a refactor of the newsletter sign up form. Currently, due
to time contraints, I was not able to dig into Coherence and remove the
requirement for a sign up to include a password. Some where deep in the
weeds was a section of code looking for a password even though I thought
I had removed it from everywhere. This led me to doing some grossness of
submitting a password in a hidden form field. Anyone could see this
password and use it to sign in on the admin page. I don't want a
non-admin sending newsletters out so there is a check to make sure a
signed in user is an admin and if not, they're told to bugger off. Not
ideal by any stretch but it gets things working for now until I can dig
in and refactor that part out.