Add tests for creating a newsletter
Addes a test for creating a newsletter. The testing for Swoosh is a
little lacking for now. The best I can do is assert that when I send a
user and the newsletter to my NewsletterEmail module, I get an email
back whose subject matches what was passed through the params. That
coupled with a simple creating one in the dev environment will have to
suffice for now.