Set default password to be chosen by env variable
In removing the Coherance function to verify a password, all passwords
were getting set to nil. This was fine as no one would use a password
except an admin (and you could manually add that). However, when writing
the acceptance test for an admin, it turns out that the nil password
could not be changed due to the Coherence function being removed.

It put that function back, and added environment variables to be used to
set a password on user sign up. This will still keep a user from logging
in as they won't see the password and allows the admin password to be
changed as needed.