Create test for an admin logging in
It turns out that this was a giant pain in the ass. When I first created
the test for an admin to log in, I kept getting back page_source that
had "incorrect email or password" instead of the expected "Create a
newsletter". It took the longest time to figure out that what was
happening was when ex_machina was inserting an admin into the database,
even though I specified a password in the factory, the password was
getting set to nil. This makes sense in hindsight as that is how things
are set up for when a user signs up for the newsletter.

For the life of me, I couldn't get the admin to have a password in the
test or to have be able to be updated with a password. Thankfully I
found https://github.com/thoughtbot/ex_machina/issues/82 which details
how to go about inserting a user with password hashing involved. It was
merely a matter of having a function in my factory that takes the user
and the password and uses Comeonin.Bcrypt to hash it and then updated
the user with the :password_hash.

There is some more tidying that could be done to user.ex and the factories but I
was so happy to have this working I wanted to get it committed.

All tests are green.