Move confirmed_at from test to factory
I was setting confirmed_at when I built the user in the test. Went ahead
and moved that into the factory as it will be used for every admin that
is created and not just the one in this test.

Also removed the :password and :password_confirmation from the user.ex
changeset cast as it was not needed.