Create acceptance test for sending a newsletter
Add one more test for an admin who signs in and then sends out a
newsletter. This involved updating a couple of buttons to have id's as
well as the form fields to have an id. Also had to change the config for
each environment as the testing environment was actually using SendGrid
to send out an email.

All tests are green.