Acceptance tests (#11)
* Setup and install Hound for acceptance testing

* Add acceptance tests for a visitor

* Set default password to be chosen by env variable

In removing the Coherance function to verify a password, all passwords
were getting set to nil. This was fine as no one would use a password
except an admin (and you could manually add that). However, when writing
the acceptance test for an admin, it turns out that the nil password
could not be changed due to the Coherence function being removed.

It put that function back, and added environment variables to be used to
set a password on user sign up. This will still keep a user from logging
in as they won't see the password and allows the admin password to be
changed as needed.

* Create test for an admin logging in

It turns out that this was a giant pain in the ass. When I first created
the test for an admin to log in, I kept getting back page_source that
had "incorrect email or password" instead of the expected "Create a
newsletter". It took the longest time to figure out that what was
happening was when ex_machina was inserting an admin into the database,
even though I specified a password in the factory, the password was
getting set to nil. This makes sense in hindsight as that is how things
are set up for when a user signs up for the newsletter.

For the life of me, I couldn't get the admin to have a password in the
test or to have be able to be updated with a password. Thankfully I
found https://github.com/thoughtbot/ex_machina/issues/82 which details
how to go about inserting a user with password hashing involved. It was
merely a matter of having a function in my factory that takes the user
and the password and uses Comeonin.Bcrypt to hash it and then updated
the user with the :password_hash.

There is some more tidying that could be done to user.ex and the factories but I
was so happy to have this working I wanted to get it committed.

* Move confirmed_at from test to factory

I was setting confirmed_at when I built the user in the test. Went ahead
and moved that into the factory as it will be used for every admin that
is created and not just the one in this test.

Also removed the :password and :password_confirmation from the user.ex
changeset cast as it was not needed.

* Create acceptance test for sending a newsletter

Add one more test for an admin who signs in and then sends out a
newsletter. This involved updating a couple of buttons to have id's as
well as the form fields to have an id. Also had to change the config for
each environment as the testing environment was actually using SendGrid
to send out an email.

All tests are green.